Add `cancel` command
Cancels a running project started by the `up` command. It simply
sends a `SIGTERM` signal to the PID of the Director instance that
started the project, so that when Director handles that signal it
will terminate the processes it started.

This is useful especially when the user makes a mistake and wants
to cancel the project, especially if the project has been running
for a long time (like a VM). So, instead of waiting for the project
to finish, the user can simply cancel its execution, fix the error
and deploy again.