Store lock files in the temporary directory
The lock files are stored on a per-project basis, which is undesirable
because it can cause that after an operating system crash or simply
unplugging the power cable the lock files remain stale, preventing
Director from running again. The current workaround is to use the
temporary directory that Director assumes is removed on boot
(`clear_tmp_enable=YES` in your `rc.conf(5)`).